backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
anthropic==0.40.0
faster-whisper==1.1.0
elevenlabs==1.9.0
python-dotenv==1.0.1
numpy==1.26.4
httpx==0.27.2
